Chatbot/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pandas
requests
python-dotenv
pydantic
